Fixing issue with broken symlinks
Problem:
When traversing a directory containig broken symbolic links, said links
throw error before we have a chance to ignore them, which we might want
to.

Solution:
When encountering error code `ENOENT` make `lstat` call which doesn't
fail on broken links due to not following them. It is up to the user
then to check if he needs to ignore this particular link via ignore
function or pattern.

Fixes #42